# openapi.yaml
openapi: 3.0.3
info:
  title: MyNotes API
  description: |
    REST API for the mynotes application. A `notes` resource demonstrates the
    full stack: OpenAPI → ogen server stubs → handler → service → repository →
    SQLite. Endpoints are served under the `/api/v1` base path (applied via the
    server URL and ogen's path prefix). Timestamps use RFC 3339 (UTC). Errors
    return `{"error": "message"}`.

    # Markdown dialect

    A note's `content` is stored and returned as verbatim Markdown. Consumers
    that render notes (the embedded web UI, the native mobile app, HTML export)
    should interpret it per the **Markdown dialect specification** in
    `markdown-spec.md`. In brief, it is [CommonMark](https://spec.commonmark.org/)
    plus a fixed subset of [GFM](https://github.github.com/gfm/) (tables,
    strikethrough, autolinks, task lists), a safe raw-HTML subset, and several
    application-specific render-time transforms — AsciiMath math (`$x^2$`), inline
    Lucide icons (`[!name]`), box / foldable blockquotes (`>-`, `>+`, `>*`) and
    the callouts built from them, internal wikilinks (`[[slug]]`), and artifact
    image references (`![alt](artifact:<sha256>)`). All
    transforms are stored verbatim; a consumer that passes `content` through
    unchanged receives the literal source. See `markdown-spec.md` for the
    full grammar, the callout alias/colour table, and the wikilink indexing rules.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
...
paths:
  /notes:
    get:
      operationId: listNotes
      summary: List notes
      description: >
        Returns a page of notes. The browse list (no `q`) is ordered by the
        `sort`/`order` parameters, defaulting to update time, newest first.
        When `q` is supplied, results are restricted to notes matching the
        full-text query and ordered by relevance (`sort`/`order` are ignored).
      parameters:
        - name: q
          in: query
          description: Optional full-text search query.
          required: false
          schema:
            type: string
            maxLength: 200
        - name: tag
          in: query
          description: >
            Optional tag slug filter. May be repeated; a note must carry ALL
            supplied tags to match (AND semantics). When combined with `q`, that
            filter also applies (AND). The single-value form (`tag=x`) is a
            one-element filter and behaves as before.
          required: false
          explode: true
          schema:
            type: array
            maxItems: 20
            items:
              type: string
              maxLength: 100
              pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
        - name: titlePrefix
          in: query
          description: >
            When true and `q` is supplied, `q` is matched as a case-insensitive
            prefix of the note title (autocomplete style) instead of a full-text
            search; content is not matched. Ignored when `q` is empty.
          required: false
          schema:
            type: boolean
            default: false
        - name: sort
          in: query
          description: >
            Field the browse list is sorted by. `updated` (default) and
            `created` sort by the respective timestamp; `title` sorts
            alphabetically, case-insensitively. Ignored for full-text search
            (`q` without `titlePrefix`), which is ordered by relevance, and for
            `titlePrefix` matching, which is always ordered by title.
          required: false
          schema:
            type: string
            enum: [updated, created, title]
            default: updated
        - name: order
          in: query
          description: >
            Direction for `sort`. `desc` (default) is newest-first for the
            timestamp fields and Z→A for `title`; `asc` is the reverse.
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: A page of notes.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NoteList"
        default:
          $ref: "#/components/responses/Error"
    post:
      operationId: createNote
      summary: Create a note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNoteRequest"
      responses:
        "201":
          description: The created note.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        default:
          $ref: "#/components/responses/Error"
  /notes/{slug}:
    parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
          maxLength: 100
          pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
    get:
      operationId: getNote
      summary: Get a note by slug
      responses:
        "200":
          description: The requested note.
          headers:
            Etag:
              description: >
                Strong entity tag reflecting the note's current version.
                Quoted integer, e.g. "1", "2".
              required: true
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        default:
          $ref: "#/components/responses/Error"
    patch:
      operationId: updateNote
      summary: Update a note (partial)
      description: Only the fields present in the request body are changed.
      parameters:
        - name: If-Match
          in: header
          required: false
          description: >
            When provided, the update is rejected with 412 Precondition Failed
            unless the note's current version matches this Etag (quoted integer).
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNoteRequest"
      responses:
        "200":
          description: The updated note.
          headers:
            Etag:
              description: Etag of the note after the update.
              required: true
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "412":
          description: >
            Precondition Failed — If-Match was provided and the current version
            does not match.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          $ref: "#/components/responses/Error"
    delete:
      operationId: deleteNote
      summary: Delete a note
      responses:
        "204":
          description: Note deleted.
        default:
          $ref: "#/components/responses/Error"
  /tags:
    get:
      operationId: listTags
      summary: List all tags
      description: >
        Returns all tags sorted by slug, each with the number of notes carrying
        it. Powers client-side autocomplete and the tag-management view.
      responses:
        "200":
          description: All tags.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagList"
        default:
          $ref: "#/components/responses/Error"
    post:
      operationId: createTag
      summary: Create a tag
      description: >
        Tags must be created explicitly before being attached to a note (see
        the `tags` field on CreateNoteRequest/UpdateNoteRequest). A tag is
        identified solely by its `slug`, which is also its display label. A
        `slug` that collides with an existing tag is a 409.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTagRequest"
      responses:
        "201":
          description: The created tag.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
        default:
          $ref: "#/components/responses/Error"
  /tags/{slug}:
    parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
          maxLength: 100
          pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
    delete:
      operationId: deleteTag
      summary: Delete a tag
      description: >
        Removes the tag and detaches it from every note that had it. No
        orphan-prevention; deleting an unknown slug is a 404.
      responses:
        "204":
          description: Tag deleted.
        default:
          $ref: "#/components/responses/Error"
  /import:
    post:
      operationId: importNote
      summary: Import a document as a note
      description: >
        Imports a document and saves it as a note. Use `Content-Type: text/html`
        to import HTML (converted to Markdown; title taken from the <title>
        element or the first heading). Use `Content-Type: text/markdown` to
        import Markdown directly (title taken from the first ATX heading or the
        filename). The produced Markdown is subject to the same validation as
        regular note creation.
      requestBody:
        required: true
        content:
          text/html:
            schema:
              type: string
          text/markdown:
            schema:
              type: string
      responses:
        "201":
          description: The created note.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        default:
          $ref: "#/components/responses/Error"
  /artifacts:
    post:
      operationId: createArtifact
      summary: Upload an artifact
      description: >
        Uploads binary content and stores it as a content-addressed artifact
        keyed by the SHA-256 of the content. Uploading the same bytes twice
        returns the existing record unchanged. Supported content types are the
        common image formats. The SHA-256 hex digest is returned in the response,
        used in the artifact URL, and is how note content references the
        artifact: `![alt](artifact:<sha256>)`, an application-defined URL scheme
        a rendering consumer resolves against wherever it fetches artifacts from
        (see `markdown-spec.md`).
      requestBody:
        required: true
        content:
          image/png:
            schema:
              type: string
              format: binary
          image/jpeg:
            schema:
              type: string
              format: binary
          image/gif:
            schema:
              type: string
              format: binary
          image/webp:
            schema:
              type: string
              format: binary
          image/svg+xml:
            schema:
              type: string
              format: binary
          application/mathml+xml:
            schema:
              type: string
              format: binary
      responses:
        "201":
          description: The created (or already existing) artifact.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artifact"
        default:
          $ref: "#/components/responses/Error"
  /artifacts/{sha256}:
    parameters:
      - name: sha256
        in: path
        required: true
        schema:
          type: string
          pattern: "^[0-9a-f]{64}$"
    get:
      operationId: getArtifact
      summary: Download an artifact by SHA-256
      description: >
        Returns the raw artifact content with its original `Content-Type`.
        The SHA-256 hex digest in the path identifies the artifact.
      responses:
        "200":
          description: The artifact content.
          content:
            image/png:
              schema:
                type: string
                format: binary
            image/jpeg:
              schema:
                type: string
                format: binary
            image/gif:
              schema:
                type: string
                format: binary
            image/webp:
              schema:
                type: string
                format: binary
            image/svg+xml:
              schema:
                type: string
                format: binary
            application/mathml+xml:
              schema:
                type: string
                format: binary
        "404":
          description: Artifact not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          $ref: "#/components/responses/Error"
    delete:
      operationId: deleteArtifact
      summary: Delete an artifact
      responses:
        "204":
          description: Artifact deleted.
        default:
          $ref: "#/components/responses/Error"
  /icons/{set}/{name}:
    parameters:
      - name: set
        in: path
        required: true
        schema:
          type: string
          maxLength: 64
          pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
      - name: name
        in: path
        required: true
        schema:
          type: string
          maxLength: 64
          pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
    get:
      operationId: getIcon
      summary: Get an icon as SVG
      description: >
        Returns the named icon from the given set as an `image/svg+xml` document.
        Icons are fixed, built-in sets; note content references them inline as
        `[!<name>]` (or as Markdown images `![<name>](/api/v1/icons/<set>/<name>)`,
        which this endpoint serves). The response is a static, public,
        immutable asset served under a locked-down sandbox Content-Security-Policy. 
        The icon has a fixed stroke colour (it is loaded via `<img>`, so it cannot 
        inherit the client's theme). Unknown sets or names return 404.
        Currently only the `lucide` set is available (from the vendored `lucide-static`, 
        see https://lucide.dev/), but others might be added in future.
      responses:
        "200":
          description: The icon SVG document.
          content:
            image/svg+xml:
              schema:
                type: string
                format: binary
        "404":
          description: Icon not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          $ref: "#/components/responses/Error"
  /notes/{slug}/download-markdown:
    parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
          maxLength: 100
          pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
    get:
      operationId: downloadNoteMarkdown
      summary: Download a note as raw Markdown
      description: >
        Returns the note `content` as a raw Markdown body with a
        `Content-Disposition` attachment header. Errors keep the JSON
        `{"error": "message"}` shape.
      responses:
        "200":
          description: The note content as raw Markdown.
          headers:
            Content-Disposition:
              description: Attachment header carrying the download filename.
              required: true
              schema:
                type: string
          content:
            text/markdown:
              schema:
                type: string
        "404":
          description: Note not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          $ref: "#/components/responses/Error"
  /notes/{slug}/publish:
    parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
          maxLength: 100
          pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
    put:
      operationId: publishNote
      summary: Publish a note as a public page
      description: >
        Stores the rendered HTML of a note and serves it, without
        authentication, at `/public/notes/{slug}`. The request body is the
        note content rendered to an HTML *fragment* — not a document; the
        server wraps it and supplies the stylesheet. The client renders it,
        because the Markdown dialect is defined by the render pipeline the
        clients share (see `markdown-spec.md`) and not by the server.

        The fragment is sanitized before it is stored: `<script>`, framed
        content, event handlers and disallowed URL schemes are removed rather
        than rejected, since this is rendered output and not a source of truth.
        Images must reference artifacts in the same `artifact:<sha256>` form
        note content uses; the server rewrites them to public artifact URLs and
        makes exactly those artifacts publicly readable. Referencing an
        artifact that is not stored is a 400.

        The published page is a snapshot. Editing the note afterwards does not
        change it — publish again to update it. Publishing an already-published
        note replaces the snapshot; deleting the note withdraws it.
      requestBody:
        required: true
        content:
          text/html:
            schema:
              type: string
              maxLength: 2000000
      responses:
        "200":
          description: The note is published.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublishedNote"
        default:
          $ref: "#/components/responses/Error"
    delete:
      operationId: unpublishNote
      summary: Withdraw a published note
      description: >
        Removes the note's public page. Artifacts referenced only by this
        published note stop being publicly readable. A note that is not
        published is a 404.
      responses:
        "204":
          description: The note is no longer published.
        default:
          $ref: "#/components/responses/Error"
  /notes/{slug}/split:
    parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
          maxLength: 100
          pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
    post:
      operationId: splitNote
      summary: Split a note into several notes by Markdown headings
      description: >
        Creates one new note per section of the source note, cutting at the
        shallowest ATX heading level present (headings inside fenced code blocks
        are ignored). Content before the first such heading is discarded. Each
        new note takes its title from the section heading and shares the source
        note's `created_at` and `updated_at`; the source note is left unchanged.
        A note with no headings is a 400. The optional `tag` must name an
        existing tag (400 if unknown), attached to every new note.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SplitNoteRequest"
      responses:
        "201":
          description: The notes created from the split, in document order.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SplitNoteResponse"
        default:
          $ref: "#/components/responses/Error"
components:
  responses:
    Error:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    NoteList:
      type: object
      required: [total, notes]
      properties:
        total:
          type: integer
          description: Number of notes returned in this page.
        notes:
          type: array
          items:
            $ref: "#/components/schemas/NoteSummary"
    CreateNoteRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        content:
          type: string
          maxLength: 1000000
          default: ""
          description: >
            Markdown content; see the Markdown dialect specification
            (markdown-spec.md). HTML carrying disallowed constructs is
            rejected (400).
        slug:
          type: string
          maxLength: 100
          pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
        tags:
          type: array
          items:
            type: string
            maxLength: 100
            pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          description: >
            Slugs of existing tags to attach. Tags must already exist (create
            via POST /tags first) — an unknown slug is a 400. Absent or
            omitted defaults to no tags.
    UpdateNoteRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        content:
          type: string
          maxLength: 1000000
          description: >
            Markdown content; see the Markdown dialect specification
            (markdown-spec.md). HTML carrying disallowed constructs is
            rejected (400).
        tags:
          type: array
          items:
            type: string
            maxLength: 100
            pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          description: >
            Slugs of existing tags. Absent leaves the note's tags unchanged;
            present (including an empty array) replaces the full set. An
            unknown slug is a 400. Replacing with the same set as before is a
            no-op (does not bump `version`).
...
    Tag:
      type: object
      required: [slug]
      properties:
        slug:
          type: string
          maxLength: 100
          pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
...
    TagList:
      type: object
      required: [tags]
      properties:
        tags:
          type: array
          items:
            $ref: "#/components/schemas/TagSummary"
    CreateTagRequest:
      type: object
      required: [slug]
      properties:
        slug:
          type: string
          maxLength: 100
          pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string